Refactor BrowserPanel to ProfilePanel and modify test
